API to filter out possible results.
Sometimes it's valuable for the checker to have other information about
the result under question, such as original sentences or their
attributes.

This commit deprecates checker API and adds a new filter API. The
difference between the two is that checker takes string as input, while
filter takes the whole result object.